LayoutTests/editing/deleting/delete-line-break-between-paragraphs-with-same-style-live-range-expected.txt

This tests deleting a line break between paragraphs with the same inline style. Inline style should be preserved after the merge.
| "\n"
| <div>
|   <font>
|     class="Apple-style-span"
|     face="monospace"
|     "hello world<#selection-caret>"
|   <span>
|     style="font-family: monospace;"
|     "WebKit"
| "\n"
